+++ packages/zod-schemas/src/url.spec.ts
@@ -1,3 +1,4 @@
+import { ZodSchemaId } from "./schema-id";
 import { createUrlSchema, UrlSchema } from "./url";
 
 describe("UrlSchema", () => {
@@ -66,6 +67,12 @@ describe("UrlSchema", () => {
       const invalidUrl = "";
 
       expect(() => UrlSchema.parse(invalidUrl)).toThrowError("Invalid URL.");
+    });
+
+    it("should have the correct schema id", () => {
+      expect(UrlSchema.meta()).toEqual({
+        id: ZodSchemaId.Url,
+      });
     });
   });